Manage suggestions dropdown directly from search-input
The same code was duplicated on EmbeddedSearch and PopupSearch, while suggestion is a feature belonging to the SearchInput.
