+++ libs/search-widget/src/old-components/embedded-search/EmbeddedSearch.scss
@@ -8,8 +8,4 @@
     padding: var(--rhythm-1) 0;
     display: none;
   }
-
-  .suggestions {
-    padding: var(--rhythm-2);
-  }
 }
