+++ libs/search-widget/src/old-components/embedded-search/EmbeddedSearch.svelte
@@ -5,8 +5,6 @@
   import Results from '../results/Results.svelte';
   import SearchInput from '../search-input/SearchInput.svelte';
   import { _ } from '../../core/i18n';
-  import Modal from '../../common/modal/Modal.svelte';
-  import Suggestions from '../suggestions/Suggestions.svelte';
 
   export let placeholder = '';
 
@@ -20,40 +18,13 @@
       inTitleOnly: checked,
       highlight: true,
     });
-
-  const paragraphs = nucliaState().paragraphs;
-  const intents = nucliaState().labelIntents;
-  const hasSearchError = nucliaState().hasSearchError;
-  let inputElement: HTMLElement | undefined;
-  let position: DOMRect | undefined;
-  let showSuggestions = false;
-
-  const openSuggestions = () => {
-    showSuggestions = true;
-    position = inputElement?.getBoundingClientRect();
-  };
-  const closeSuggestions = () => {
-    showSuggestions = false;
-  };
 </script>
 
 <div class="sw-embedded-search">
   <div class="search-input-container">
-    <div bind:this={inputElement}>
-      <SearchInput embeddedSearch={true} {placeholder} on:typeahead={openSuggestions} on:search={closeSuggestions} />
-    </div>
+    <SearchInput embeddedSearch={true} {placeholder}/>
   </div>
-  <Modal
-    show={showSuggestions && ($paragraphs.length > 0 || $hasSearchError)}
-    on:close={closeSuggestions}
-    popup={true}
-    parentPosition={position}
-    --modal-width-md="var(--search-bar-max-width)"
-  >
-    <div class="suggestions">
-      <Suggestions paragraphs={$paragraphs} intents={$intents} />
-    </div>
-  </Modal>
+
   <div class="options">
     <Toggle label={$_('form.title-only')} on:change={(e) => onChange(e.detail)} />
   </div>
