+++ libs/search-widget/src/old-components/popup-search/PopupSearch.scss
@@ -1,8 +1,4 @@
 .sw-popup-search {
-  .suggestions {
-    padding: var(--rhythm-2);
-  }
-
   .results {
     max-width: 100vw;
     width: var(--default-modal-width);
